React Challenge: Lifting state up and function components
Challenge 1: Inside the JS section, create a class component named App. Inside its render() method, have it return the Welcome component. In the ReactDOM.render() method, change Welcome to App.

Challenge 2: Lift the state from the Welcome component to App, so that the state is initialized inside of App's constructor.

Challenge 3: Convert the Welcome component to a function component that returns the same welcome message as before. You will need to pass the bootcampName property of state from App to the Welcome component. It's up to you whether or not to destructure it -- either is fine. Try it both ways for practice!