Code Challenge: Ternary Operator
https://codepen.io/mbonzi/pen/bGrypQN?editors=0011

Code Challenge
There are four sets of if/else statements in the JS section of this Codepen. Refactor (rewrite) each one so that it uses the ternary operator shorthand. To get you started, we'll give you an answer to #1 for free:
console.log(1+1===2 ? 'It is true that 1+1 equals 2.' : 'It is false that 1+1 equals 2.');
That should help you figure out #2 and #3. #4 is a little different - see below.

About Refactor Challenge #4
- For this challenge, you will refactor the code inside the function isEven. You will use the ternary operator to determine what value (true or false) to return from the function. It is possible to write this as a single-line arrow function that does not require curly braces nor a 'return' keyword.
- In the helper code at line 37, there are two examples of a ternary operator used inline. You don't need to do anything with these examples -- they are just there as demos, to show you more ways to use the ?: syntax.